Use accelerometer when calibrating magnetometer
The accelerometer's measured direction of Earth can be used instead
of the "rough calibration" which is an estimate of the magnetic null
measurement based on the maximal bounds of values felt by the magnetometer.

This allows calibration to begin immediately rather than only after bounds
on all 3 axis have been established. On the downside, the device must now
be moved in some complicated motion in air, such as figure-of-8 pattern.